chore: updates dashboard,  update readme, clean samples
